+++ src/logic/angular/generateNewComponent.ts
