+++ src/logic/angular/isAngularCliAvailable.ts
