+++ src/logic/index.ts
@@ -1,5 +1,3 @@
 export * from "./generateNewComponent";
 export * from "./getChanges";
-export * from "./angularTemplateHandler";
-export * from "./ts/tsComponentHandler";
 export * from "./isAngularCliAvailable";
