+++ src/logic/ts/ast/ast-utils.ts
