+++ src/logic/ts/ast/component-property.ts
