+++ src/logic/ts/tsComponentHandler.ts
